requirements.txt
<<<<<<< HEAD
requests>=2.25.0
python-dotenv>=0.15.0
colorama>=0.4.4
rich>=10.0.0
prompt-toolkit>=3.0.20
pytest>=7.0.0  # Para testes 
=======
requests>=2.28.0
python-dotenv>=1.0.0
typer>=0.9.0
rich>=10.0.0
websockets>=12.0
prompt_toolkit>=3.0.0
mcp[cli]>=0.1.0
>>>>>>> feature/arcee-api
